requirements.txt
pwn==1.0
pwntools==4.12.0
pyserial==3.5
